+++ config/routes.rb
@@ -65,6 +65,7 @@ Rails.application.routes.draw do
     resources :agents, only: %i[create]
     resources :cases, only: %i[index show edit update] do
       resources :interactions, only: %i[new create]
+      resource :resolution, only: %i[new create], module: :cases
     end
     resources :schools, only: %i[show]
   end